backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.1
celery[redis]==5.4.0
redis==5.2.1
openpyxl==3.1.5
aiofiles==24.1.0
google-generativeai==0.8.4
websockets==14.1
